set_disk_partition_fstype:umount before mkfs,handle varitey fstype

+++ src/part_util.py
@@ -758,9 +758,20 @@ class PartUtil:
         if fstype==None or len(fstype)==0:
             print "no filesystem specified"
             return
-
+        #umount a partition before format it to make a filesystem
+        self.set_disk_partition_umount(partition)
         part_path=partition.path
-        format_command="sudo mkfs -t "+fstype+" "+part_path
+        if fstype=="xfs":
+            format_command=""
+        elif fstype=="reiserfs":
+            format_command=""
+        elif fstype=="swap":
+            format_command=""
+        elif fstype in ["ext2","ext3","ext4"]:
+            format_command="sudo mkfs -t "+fstype+" "+part_path
+        else:
+            print "invalid fstype"
+            
         run_os_command(format_command)
     
 
@@ -805,7 +816,8 @@ class PartUtil:
             else:
                 run_os_command(umount_command)
         else:
-            print "don't need to umount"+part_path
+            # print "don't need to umount"+part_path
+            return
 
     def get_disk_partition_mount(self,partition):
         '''get partition mount info,need consider multiple mount '''
